requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.10.0
pydantic-settings==2.7.0
PyMySQL==1.1.1
python-jose[cryptography]==3.3.0
httpx==0.28.0
cachetools==5.5.0
python-dotenv==1.0.1
orjson==3.10.0
structlog==24.4.0
pytest==8.3.0
pytest-asyncio==0.24.0
